Pull members up to GameState

+++ src/RoadCaptain/GameStates/OnSegmentState.cs
@@ -4,7 +4,7 @@
     {
         public Segment CurrentSegment { get; }
 
-        public OnSegmentState(int activityId, TrackPoint currentPosition, Segment segment) 
+        public OnSegmentState(ulong activityId, TrackPoint currentPosition, Segment segment) 
             : base(activityId, currentPosition)
         {
             CurrentSegment = segment;
